Update image URIs too when reindexing
Previously, if a specific buildpack-version combination existed in the
index already, all but the `addr` and `created_at` fields would be
updated if they had changed.

Now the `addr` field is updated too, leaving only the `created_at`
field as the only one not updated.

This ensures that the changes in buildpacks/registry-index#5767 are
correctly picked up.

Signed-off-by: Ed Morley <501702+edmorley@users.noreply.github.com>